Rename the command and the description
